requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
alembic
pydantic[email] 
argon2_cffi
google-generativeai
python-dotenv
requests